Move ConstantFolding out of core
