+++ docs/reference/esql/functions/types/mv_zip.asciidoc
@@ -6,5 +6,15 @@
 |===
 string1 | string2 | delim | result
 keyword | keyword | keyword | keyword
+keyword | keyword | text | keyword
+keyword | keyword | | keyword
+keyword | text | keyword | keyword
+keyword | text | text | keyword
+keyword | text | | keyword
+text | keyword | keyword | keyword
+text | keyword | text | keyword
+text | keyword | | keyword
+text | text | keyword | keyword
 text | text | text | keyword
+text | text | | keyword
 |===
